blog/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.2
certifi==2019.6.16
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-etc==1.2.0
django-extensions==3.0.2
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-richtextfield==1.6
djangorestframework==3.11.0
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
graphviz==0.14
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
isort==5.5.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.20
numpy==1.19.2
opencv-python==4.4.0.44
openpyxl==3.0.5
Pillow==7.1.2
pipenv==2020.11.15
playsound==1.2.2
protobuf==3.12.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
PyHamcrest==2.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-decouple==3.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
shell==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.1.2
